move code to get the networks into its own file to cut down on code duplication
