+++ src/command/game/scoreboard.js
@@ -2,6 +2,8 @@ import format from 'date-fns/format';
 import emoji from 'node-emoji';
 import { center } from 'wide-align';
 
+import getBroadcastNetworks from './network';
+
 import { bold, nbaRed, neonGreen } from '../../utils/log';
 import { basicTable } from '../../utils/table';
 
@@ -66,25 +68,7 @@ const scoreboard = (
   const homeTeamStartingPlayers = getStartingPlayers(homeTeam);
   const visitorTeamStartingPlayers = getStartingPlayers(visitorTeam);
 
-  const televisionNetworks = broadcasters.tv.broadcaster;
-  let homeTeamNetwork = televisionNetworks.filter(
-    broadcaster => broadcaster.home_visitor === 'home'
-  );
-  let visitorTeamNetwork = televisionNetworks.filter(
-    broadcaster => broadcaster.home_visitor === 'visitor'
-  );
-  let nationalNetwork = televisionNetworks.filter(
-    broadcaster => broadcaster.home_visitor === 'natl'
-  );
-  nationalNetwork = !nationalNetwork.length
-    ? 'N/A'
-    : nationalNetwork[0].display_name;
-  homeTeamNetwork = !homeTeamNetwork.length
-    ? nationalNetwork
-    : homeTeamNetwork[0].display_name;
-  visitorTeamNetwork = !visitorTeamNetwork.length
-    ? nationalNetwork
-    : visitorTeamNetwork[0].display_name;
+  const networks = getBroadcastNetworks(broadcasters.tv.broadcaster);
 
   scoreboardTable.push(
     vAlignCenter([
@@ -248,7 +232,7 @@ const scoreboard = (
       {
         colSpan: 10,
         content: bold(
-          `${homeTeamNetwork} ${emoji.get('tv')}  ${visitorTeamNetwork}`
+          `${networks.homeTeam} ${emoji.get('tv')}  ${networks.visitorTeam}`
         ),
         hAlign: 'center',
       },
